requirements.txt
absl-py==2.1.0
astor==0.8.1
charset-normalizer==3.4.2
coloredlogs==15.0.1
cycler==0.11.0
flatbuffers==25.2.10
fonttools==4.38.0
gast==0.2.2
google-pasta==0.2.0
grpcio==1.62.3
h5py==2.10.0
humanfriendly==10.0
idna==3.10
importlib-metadata==6.7.0
joblib==1.3.2
Keras==2.2.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
Markdown==3.4.4
MarkupSafe==2.1.5
matplotlib==3.5.3
mpmath==1.3.0
netron==8.3.5
numpy==1.18.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
onnx==1.14.1
onnxruntime==1.10.0
opencv-python==4.11.0.86
opt-einsum==3.3.0
packaging==24.0
pandas==1.3.5
Pillow==9.5.0
protobuf==3.20.3
psutil==7.0.0
py-cpuinfo==9.0.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.1
requests==2.31.0
scikit-learn==0.21.3
scipy==1.7.3
seaborn==0.12.2
six==1.17.0
sympy==1.10.1
tensorboard==1.15.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.5
termcolor==2.3.0
tf2onnx==1.16.1
threadpoolctl==3.1.0
torch==1.13.1
torchvision==0.14.1
tqdm==4.67.1
typing_extensions==4.7.1
ultralytics==8.0.145
urllib3==2.0.7
Werkzeug==2.2.3
wrapt==1.16.0
zipp==3.15.0
